remove gil from cython cyk and add openmp support.  make sure alignmenttable not lost in viterbi interface.

+++ tests/cfgTest.py
@@ -112,12 +112,12 @@ class TestCase(TestBase):
         cfgProb, cfgStates = cfg.decode(obs)
         assert_array_equal(cfgStates, [2,0,0,2])
 
-        alignment = np.array([[1],[0],[0],[1]], dtype=np.uint8)
+        alignment = np.array([[1],[0],[0],[1]], dtype=np.uint16)
         cfgProb, cfgStates = cfg.decode(obs, alignmentTrack = alignment,
                                         defAlignmentSymbol=0)
         assert_array_equal(cfgStates, [1,0,0,1])
 
-        alignment = np.array([[1],[0],[0],[2]], dtype=np.uint8)
+        alignment = np.array([[1],[0],[0],[2]], dtype=np.uint16)
         cfgProb, cfgStates = cfg.decode(obs, alignmentTrack = alignment,
                                         defAlignmentSymbol=0)
         assert_array_equal(cfgStates, [2,0,0,2])
